requirements.txt
streamlit
scikit-learn
scikit-image
joblib
numpy
pillow